libweston/matrix: Fix narrow conversion with C++
Fixes the following C++ narrow conversion:

/include/libweston-14/libweston/matrix.h: In function ‘weston_coord
weston_coord_truncate(weston_coord)’:
/include/libweston-14/libweston/matrix.h:212:39: error: narrowing
conversion of ‘(int)in.weston_coord::x’ from ‘int’ to ‘double’
[-Werror=narrowing] 212 |         return (struct weston_coord){
	(int)in.x, (int)in.y };
      |                                       ^~~~~~~~~
/home/mvlad/install-new/include/libweston-14/libweston/matrix.h:212:50:
error: narrowing conversion of ‘(int)in.weston_coord::y’ from ‘int’ to
‘double’ [-Werror=narrowing] 212 |         return (struct weston_coord){
	(int)in.x, (int)in.y };

(cherry picked from commit a501acd291e819d00ef744ac6cd8b25500b07c3c)

+++ include/libweston/matrix.h
@@ -209,7 +209,7 @@ weston_coord_surface_sub(struct weston_coord_surface a,
 static inline struct weston_coord __attribute__ ((warn_unused_result))
 weston_coord_truncate(struct weston_coord in)
 {
-	return (struct weston_coord){ (int)in.x, (int)in.y };
+	return (struct weston_coord){ (double)(int) in.x, (double)(int) in.y };
 }
 
 static inline struct weston_coord_surface __attribute__ ((warn_unused_result))
